migrate library to apps router
