+++ packages/landingpage/app/[lang]/library/patterns/components/preview-box.module.scss
