+++ packages/landingpage/utils/hooks/useStorybookUrlSyncer.ts
@@ -1,4 +1,4 @@
-import { useRouter } from 'next/router';
+import { useRouter, usePathname, useSearchParams } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import { inDevEnvironment } from '../in-dev-mode';
 import { useMount } from 'react-use';
@@ -17,7 +17,10 @@ const POST_CURRENT_STORY_TYPE = 'post-current-story';
  * This plugin sends the id of the currently displayed story and also syncs the story-id to our landingpage URL.
  */
 export const useStorybookUrlSyncer = () => {
-  const { push, query } = useRouter();
+  const router = useRouter();
+  const pathName = usePathname();
+  const searchParams = useSearchParams();
+
   const [storyId, setStoryId] = useState<string | null>(null);
 
   useMount(() => {
@@ -39,11 +42,11 @@ export const useStorybookUrlSyncer = () => {
 
   // push story changes to url
   useEffect(() => {
-    if (!storyId || !query.lang) return;
+    if (!storyId || searchParams === null) return;
 
-    push({ query: { ...query, element: storyId } }, undefined, {
-      shallow: true,
-    });
+    const query = new URLSearchParams(searchParams.toString())
+    query.set('element', storyId);
+    router.push(`${pathName}?${query}`);
   }, [storyId]);
 
   return storyId;
